Alpha feedback addressed for Beta candidate

+++ Morsel/Morsel/Source/ViewControllers/MRSLDetailCommentsViewController.m
@@ -22,7 +22,7 @@ UITableViewDelegate,
 NSFetchedResultsControllerDelegate,
 CommentTableViewCellDelegate>
 
-@property (nonatomic) int commentCount;
+@property (nonatomic) NSInteger commentCount;
 
 @property (weak, nonatomic) IBOutlet UITableView *commentsTableView;
